% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcox-package.R
\docType{package}
\name{netcox-package}
\alias{netcox-package}
\title{netcox}
\description{
Efficient procedures for fitting and cross-validating the overlapping group Lasso (implemented in C++) for Cox models with time-dependent covariates. The penalty term is a weighted sum of infinity norms of (overlapping) groups of coefficients, which can select variables structurally with a specific grouping structure.
}
